+++ Kernel/Arch/aarch64/MainIdRegister.cpp
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <Kernel/Prekernel/Arch/aarch64/MainIdRegister.h>
+#include <Kernel/Arch/aarch64/MainIdRegister.h>
 
 namespace Prekernel {
