+++ Kernel/Arch/aarch64/SerenityLogoRGB.ppm
