+++ Kernel/Arch/aarch64/linker.ld
