+++ Kernel/Arch/aarch64/vector_table.S
